Move notification mappers to their own package
