+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/mapper/notificationrequest/ManageHearingsNotificationRequestMapper.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.mapper;
+package uk.gov.hmcts.reform.finrem.caseorchestration.mapper.notificationrequest;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
